Implement login and signup functionality

+++ src/app/api/expenses/[Id]/route.ts
@@ -1,6 +1,8 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import { Prisma } from "@prisma/client";
+import {getServerSession} from "next-auth";
+import {authOptions} from "@/src/app/api/auth/[...nextauth]/route";
 
 type Params = {
   params: { id: string };
@@ -12,7 +14,11 @@ export async function DELETE(_req: Request, context: Params) {
 
   const {id} = await context.params;
   try {
-    await prisma.expense.delete({ where: { id } });
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.id) return NextResponse.json({error: "Unauthorized"}, {status: 401});
+
+    await prisma.expense.delete({ where: { id, userId: session.user.id } }); // ensures ownership
+
     return NextResponse.json({ message: "Deleted sucessfully" });
   } catch (error) {
     if (
@@ -31,11 +37,13 @@ export async function DELETE(_req: Request, context: Params) {
 
 export async function GET(_req: Request, context: {params: Promise<{id: string}>}) {
     const {id} = await context.params;
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.id) return NextResponse.json({error: "Unauthorized"}, {status: 401});
     console.log("GET request for expense id:", id)
 
   try {
     const expense = await prisma.expense.findUnique({
-      where: { id },
+      where: { id, userId: session.user.id },
     });
     if (!expense) {
       return NextResponse.json({ error: "Expense not found" }, { status: 404 });
@@ -52,11 +60,20 @@ export async function GET(_req: Request, context: {params: Promise<{id: string}>
 
 export async function PATCH(_req: Request, context: {params: Promise<{id: string}>}) {
     const {id} = await context.params;
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.id) return NextResponse.json({error: "Unauthorized"}, {status: 401});
     console.log("PATCH request for id: ", id);
 
     try {
         const data = await _req.json();
         const {title, amount, category, date} = data;
+
+        // ensure user owns this expense
+        const exp = await prisma.expense.findFirst({
+          where: {id, userId: session.user.id},
+        });
+        if (!exp) return NextResponse.json({error: "Not found"}, {status: 404});
+
         const updatedExpense = await prisma.expense.update({
             where: {id},
             data: {
@@ -67,6 +84,7 @@ export async function PATCH(_req: Request, context: {params: Promise<{id: string
             },
         });
         console.log("Expense updated: ", updatedExpense);
+        
         return NextResponse.json(updatedExpense);
     } catch (error) {
         if (
